+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Internal
 
 * `[]` caller method is extracted from `BaseDSL`, `Inverter`, and `GuardDSL` to `DSL::Caller` (nepalez)
+* `GuardDSL` is splitted to `DSL::Caller` and `DSL::Attribute` (nepalez)
 
 [Compare v0.2.0...HEAD](https://github.com/nepalez/assertion/compare/v0.2.0...HEAD)